market-brosur-sistemi/requirements.txt
# Core Flask
Flask==3.0.2
Werkzeug==3.0.1
gunicorn==21.2.0

# Security
Flask-Limiter==3.5.0
Flask-WTF==1.2.1
bleach==6.1.0

# CORS
Flask-CORS==4.0.0

# Data processing
requests==2.31.0
pillow==10.2.0
openpyxl==3.1.2
pandas==2.1.4

# Image processing (background removal)
rembg==2.0.57
onnxruntime==1.19.2

# Utils
python-dotenv==1.0.0

# OpenAI (optional)
openai==1.12.0

# Web Search (free, no API key needed)
duckduckgo_search==6.3.0
beautifulsoup4==4.12.3

# System monitoring (optional)
psutil==5.9.8
